syntropy::Byte is not an alias for standard's std::byte. Updating existing enum class to use Int instead of Byte. Adding utility cast methods.

+++ Syntropy/include/syntropy/diagnostics/severity.h
@@ -9,6 +9,8 @@
 #include <cstdint>
 #include <ostream>
 
+#include "syntropy/core/types.h"
+
 namespace syntropy
 {
     /************************************************************************/
@@ -17,7 +19,7 @@ namespace syntropy
 
     /// \brief Represents the severity of an event.
     /// \author Raffaele D. Facendola - November 2016
-    enum class Severity : Byte
+    enum class Severity : Int
     {
         /// \brief Severity used for debugging and tracing messages.
         kDebug = -1,
